scripts/pyproject.toml
[project]
name = "karp-republic-chunker"
version = "0.1.0"
description = "Scripts to chunk markdown files and upload embeddings to Neon database"
requires-python = ">=3.9"
dependencies = [
    "psycopg2-binary==2.9.7",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.2",
    "markdown>=3.5.1",
    "tiktoken>=0.5.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [] 